Reworked consumer.py and app.js to Authenicate users, Game Join Functionality, and Data Persistence
On game creation anyone who visits the the game page will be presented with the ability to join, on a join event being called, the game is no longer open and the two players are locked in.

on any event that alters the game the server unpack that data and write to the database, that data is then authenticated and serialized. After serialization we send that new game data to all connected to the channel group.

known bugs, on reconnection the game in the database is one move behind